+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+-   Add GitHub Enterprise integration
+-   Add assumptions to VCS type if the hostname matches "github", "gitlab" or "gitea"
+
 ## [0.2.0] - 2023-05-29
 
 -   Add Gitea integration